Fix Missing count on Terms panel
The Terms aggregation on Elasticsearch 5 does not contain the `missing` field and so it was hardcoded the value "unknown". To overcome this problem, we use the MissingAggregation (https://www.elastic.co/guide/en/elasticsearch/reference/5.0/search-aggregations-bucket-missing-aggregation.html) that produces the same result with the disadvantage of being another aggregation over the data.

The new aggregation is used to display the number of documents where the field is missing or null (according to configuration of the null value).